:recycle: Refactored all non-zgwclient classes into JSONClient base class
