+++ src/bptl/work_units/brp/client.py
@@ -2,13 +2,9 @@
 Implements a BRP client.
 """
 import logging
-from typing import Dict
-from urllib.parse import urljoin
 
 from django.utils.translation import gettext_lazy as _
 
-import requests
-from timeline_logger.models import TimelineLog
 from zgw_consumers.constants import APITypes
 from zgw_consumers.models import Service
 
@@ -16,6 +12,8 @@ from bptl.credentials.api import get_credentials
 from bptl.tasks.models import BaseTask, DefaultService
 from bptl.tasks.registry import register
 from bptl.work_units.zgw.tasks.base import NoService
+
+from ..clients import JSONClient
 
 logger = logging.getLogger(__name__)
 
@@ -57,43 +55,5 @@ def get_client(task: BaseTask) -> "BRPClient":
     return client
 
 
-class BRPClient:
-    task = None
-
-    def __init__(self, service: Service, auth_header: Dict[str, str]):
-        self.api_root = service.api_root
-        self.auth = auth_header
-
-    def get(self, path: str, *args, **kwargs):
-        url = urljoin(self.api_root, path)
-
-        # add the API headers
-        headers = kwargs.pop("headers", {})
-        headers.update(self.auth)
-        kwargs["headers"] = headers
-        kwargs["hooks"] = {"response": self.log}
-
-        response = requests.get(url, *args, **kwargs)
-        response.raise_for_status()
-
-        return response.json()
-
-    def log(self, resp, *args, **kwargs):
-        response_data = resp.json() if resp.content else None
-
-        extra_data = {
-            "service_base_url": self.api_root,
-            "request": {
-                "url": resp.url,
-                "method": resp.request.method,
-                "headers": dict(resp.request.headers),
-                "data": resp.request.body,
-                "params": resp.request.qs,
-            },
-            "response": {
-                "status": resp.status_code,
-                "headers": dict(resp.headers),
-                "data": response_data,
-            },
-        }
-        TimelineLog.objects.create(content_object=self.task, extra_data=extra_data)
+class BRPClient(JSONClient):
+    pass
